Fix to allow syncing of arbitrary markdown
The solution is as follows:

When syncing *to* trello we use the indent of the first line as the
indent for the entire description.

When syncing *from* trello we add a fixed indent of 2 to the
description.  This prevents markdown syntax from being interpreted as
org-mode syntax.

Somewhat a workaround for issue #89 the markup isn't translated, but at
least it is preserved