Draft complete (without "How To" section). Separate config. variable headings with indices. Added many concept indices.
This draft covers all sections of the web site, EXCEPT "How-To" and "Videos". I moved the videos to a link under "About" because the videos will never work in the PDF and INFO formats and would suffer in the HTML format, as it is constrained by the Texinfo filter. Suggest that the web site have two branches: videos and manual.
I have tenatively left out the "How To" section as the "Commands" format covers much of the same ground AND the material appears to be outdated. I think there is a place for a "How To" ... "Work with Cards", "Work with CheckLists", etc. but it will take more work to recreate those sections accurate to the current version.
You may review content (words) in either org, info, html (prettiest) or pdf. PDF may have the best comment tools for detailed review. There are also format issues, but those are after the content. I will also review later. Too soon now. -- Daniel Birket